backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.10
pydantic==2.6.0
python-dotenv==1.0.0
google-generativeai==0.4.0
python-multipart==0.0.6
starlette==0.37.0

